Improve downloading of OSCAL output templates
See issue https://github.com/GovReady/govready-q/issues/1018
for additional information.

Improve the UI on the task-finish page to display
option to download OSCAL versions of SSP.

First, create new display of generated documents listed
in a table format with different download options.
This allows for only showing option to download OSCAL version
when template is strictly for OSCAL. This display also makes
it clearer what documents have been generated than when every
document is in its own accordian. (When each output document
was displayed within an accordian, it is not clear to user
what documents are available.) Still showing accordians at this
time, but they are lower on the page.