Add Project methods to upgrade newer version of root_task app
Add new Project methods to make it possible to upgrade to a
newever version of the project's root_task app after the app
has been loaded into the database. These new methods are similar
to the admin command `upgrade_project`, but are changed slightly
to work better in the context of a Project object methods.