Merge master and 0.9.0.dev-wazuh-experimental-2019-10-22
This commit merges and removes conflicts between master (11/10/2020)
and 0.9.0.dev-wazuh-experimental-2019-10-22 where experimental
work was done integrating data from Wazuh agents into GovReady-Q.

The experimental Wazuh branch established new models in `itsystems`
to represent instances of systems in different environemnts such
as dev, test, staging, prod. This allowed Hosts to be associated
with an instance of the IT system and Agents that were observing
the Hosts.

Agents were associated with individual Agent Services, such as Wazuh.

Summerizing, the data model was extended to represent instances of
an IT System that was deployed on actual hosts and identified Agents
(with unique IDs) installed on the hosts that belong to an AgentService.

This would allow the GovReady code to use specific APIs for a given
AgentService to fetch information from the agents regarding the specific
Hosts.

There may be better approaches for this integration.