Create a new sample compliance app that contains a simple questionnaire and Word output template q-files/vendors/govready/govready-q-files-startpack/q-files/ice-cream-docx-demo It is inside such a compliance app the word template would live.
Also add a new output format to handle the docxtpl format so module_logic
and other scripts would correctly handle the format. Because module_logic relies
on metadata to define the template format, and because we normally stick that
metadata at the top of markdown file which holds the text for the template,
we need to modify the output document file to contain the metadata including
a new key called file that indicates the relative path to the docx file.
(This is mocked up the ice_cream_order_docxtpl.md file but, GovReady
does not yet properly understand how to file the value to the relative docx file.)

Update templates/task-finished.html to display a link to the new download format.
Fortunately, this is working using a slightly modified version of the initial
integration work, demonstrating we can access that template format
(once it is properly built elsewhere). (NOTE: the hack includes some hard coding
unique to local workstation file path.)

To do: Make sure that when guidedmodules.app_loading.py runs, that the docx template
is stored (and eventually retrieved) from the database. This is necessary because all
compliance apps are loaded into the database so as to not need to access anything
from the file system.