Improve statement vs cerftified statement
Make code clearer in the code that functions update
the 'smt_prototype' (e.g., statements of type control_implementation_prototype)
Continue to refer to upstream smt_prototype as "certified" statement in the UI.

Make improvements in the UI to display if statement differs and
if the statements are the same.

Improve the updating of local statement from stm_prototype to multiline statements.
Previously, a multiline statement was breaking the inline jQuery updates.
Move the updating of local statement to a distinct function to better handle
multiline statements and other details.

Better toggle displayed of certified smt_prototype statements.