Override default allauth account templates
Override default allauth account templates for login, signup,
and password reset. Override templates are in templates/account
and incorporate tags from `../venv/lib/python3.8/site-packages/allauth/`
templates.

This replaces the initial fix for styling login and sign up in
v0.9.1.46.4 (commit hash 2b4d17fb, PR #1092) that used routing
to redirect to home page's sign up and login forms. It is necessary
to preserve routing to allauth's forms because those forms
support "NEXT" URL value which is necessary for handling signup and
login for invitations to particular projects and routing person
to the correct project upon login.

This also fixed a test that was failing due to the routing from
accepting and invitation.