Add all available control implementation statements of a component to a system
even for controls that are not selected controls.

This is done on the viewpoint that the component associated
with a system should reflect the component's potential and it
is the responsibility of the system to filter the relevant information
based on the selected controls (and other customizations). This
approach is similar to Git having a local representation of all
branches and history even if the developer is working with the
perspective (or context) of a single branch.