Add Deployment model, route, etc to track system inventory
Add controls.Deployment object, related routes, views, templates,
and admin to track system deployments and deployment inventory items.
Add DeploymentForm for Deployment model. Consolidate view for creating
and editing deployments into single view.

Add simple-history to Deployment records in order to maintain an inventory
history for each deployment.

One system has multiple deployments (e.g., dev, stage, prod).
Each deployment contains an inventory of the actual endpoints/items
in a deployment of the system.

The "design" deployment by convention is a special deployment to represent
the system architecture.

Populate every new system with default deployments design, dev, stage, prod.

Deployment.inventory_items are represented as JSON data object
following a scheme that is similar to OSCAL inventory-item section.
Data for deployment inventory-items is assumed to be generated outside of GovReady.
It is critical that the inventory items have UUIDs prior to import.
Inventory item UUIDs for the life of the instantiated inventory item.

Inventory items in an deployment can be associated with an inventory item
in the "design" deployment by referencing the "design" inventory item's UUID.
This enablea a virtual persistence of an inventory-item across different
instances of the "same" assest, such as a virtual database server.

An inventory is currently added to a deployment by copying and pasting
the JSON object representing the deployment.

TODO
- Check user has permission to view project to view/create deployments and view history
- Write tests
- API for pushing data and updating inventory
- Manual upload of inventory