[WIP] Add System Assessment Results tracking.
Added `SystemAssessmentResult` model, views, forms, routes, and history
to controls to store JSON collection of assessments associated with a
system.

SystemAssessmentResults will hold a JSON collection of individual
system assessment results. The JSON collection will contain summary
information from each assessment to represent in a dashboard and
provide links to the full assessment results that may reside elsewhere
for security reasons.

SystemAssessmentResults (or sars), are primarily related to the system
with an optional relation to a system deployment.

Adds:
- Added test data for inventory items and test assessment results

Changes:
- Add `Assessments` button to action buttons on project page
- Move the `Deployment` button further down in action buttons

Fixes:
- Inlcude ProjectForm in context for Deployment and SystemAssessmentResult
  pages to allow proper starting of new projects from those pages.

Todos:
- Write tests for SystemAssessmentResult and Deployments