Project mini-dashboard now reports controls "addressed"
Trying to report controls implemented is a bit to tricky
at the moment because maybe controls need to be judged as
"implemented" by an assessor. Instead, refer to controls
addressed. Calculate "addressed" controls by looking at
consumed statements of type control_implementation for a
system root element and do a distinct to only count controls
once.